+++ python/mxnet/gluon/__init__.py
@@ -18,4 +18,4 @@ from . import utils
 
 from . import data
 
-from . import models
+from . import model_zoo
