+++ python/mxnet/gluon/model_zoo/__init__.py
@@ -1,0 +1,6 @@
+# coding: utf-8
+"""Predefined and pretrained models."""
+
+from . import model_store
+
+from . import vision
