+++ python/mxnet/gluon/model_zoo/vision/__init__.py
@@ -29,9 +29,12 @@ i.e. mini-batches of 3-channel RGB images of shape (N x 3 x H x W),
 where N is the batch size, and H and W are expected to be at least 224.
 The images have to be loaded in to a range of [0, 1] and then normalized
 using ``mean = [0.485, 0.456, 0.406]`` and ``std = [0.229, 0.224, 0.225]``.
-The transformation should preferrably happen at preprocessing. Alternatively,
-you can use the following transform to normalize::
-    normalize = mxnet.gluon.models.ColorNormalize()
+The transformation should preferrably happen at preprocessing. You can use
+``mx.image.color_normalize`` for such transformation::
+    image = image/255
+    normalized = mx.image.color_normalize(image,
+                                          mean=mx.nd.array([0.485, 0.456, 0.406]),
+                                          std=mx.nd.array([0.229, 0.224, 0.225]))
 
 .. _AlexNet: https://arxiv.org/abs/1404.5997
 .. _DenseNet: https://arxiv.org/abs/1608.06993
@@ -53,8 +56,6 @@ from .resnet import *
 from .squeezenet import *
 
 from .vgg import *
-
-from .utils import *
 
 def get_model(name, **kwargs):
     """Returns a pre-defined model by name
